Fix Sales invoice decimal drift, add Invoice Preview/Print, add expiry batch tracking
1. Sales & Billing decimal bug: Invoice.totalAmount, InvoiceItem.quantity/
   unitPrice/subtotal, Customer.creditLimit/outstandingBalance, and
   CreditCollection.amountCollected were all mapped through Java Double even
   though their MySQL columns are DECIMAL(10,2). Binary floating-point can't
   represent most 2-decimal values exactly, which is what was landing
   invoice totals a cent off from what was actually entered. Switched all of
   them to BigDecimal end to end (entities, SalesService arithmetic with
   explicit setScale(2, HALF_UP), request/response DTOs) so the value
   entered is the exact value stored and displayed. Verified with curl:
   qty 1 x Rs.100.00 -> totalAmount 100.00 exactly, and qty 3 x Rs.19.99 (a
   classic float-precision trap) -> 59.97 exactly.

   Also found and fixed a second passwordHash leak while doing this:
   POST /api/sales/invoice was returning the raw Invoice entity, whose
   nested salesOfficer (AppUser) serialized passwordHash straight into the
   response - same bug class as the /api/sales/invoices leak fixed
   previously. It now returns the same safe InvoiceDetail projection the
   Invoice Preview screen uses.

2. Invoice Preview + Print: added GET /api/sales/invoice/{id} (customer,
   officer, line items, totals - safe projection, no nested AppUser) and a
   new InvoicePreviewView - a proper invoice document (CSI logo, invoice #,
   date, customer, itemized products/qty/price/subtotal, total) opened from
   the Invoice List's new "View" button, and automatically right after an
   invoice is created instead of leaving a one-line success message as the
   only trace. Print button uses javafx.print.PrinterJob for a real print
   attempt - shows "No printer is currently connected" when
   createPrinterJob() finds none, and a clear error if the print itself
   fails, rather than being decorative.

3. Expiry date tracking (Inventory / Production / QC), per spec:
   - erp_backend.util.ExpiryCalculator: Milk Packets -> +2 months, Yogurt ->
     +3 months, everything else -> +1 year, computed automatically from a
     production/stock-in date - never entered by hand. Mirrored on the
     frontend (ExpiryRules) for a live preview only; the backend's value is
     always the one actually saved.
   - ExpiryBatch gains ProductionDate (migration_add_expiry_batch_
     production_date.sql - manual, ddl-auto=none, same as seed_roles.sql).
     Each finished-product Stock In movement now creates its own batch with
     its own production date and auto-calculated expiry date, rather than
     one fixed expiry per product - two runs of the same product on
     different dates get two batches, each tracked separately.
   - Stock leaving a product (a Stock Out movement, a sale, or a damaged-
     stock record) depletes ExpiryBatch rows FIFO (soonest-expiring batch
     first, matching how a warehouse actually rotates stock) so "available
     quantity" per batch stays accurate wherever stock goes.
   - New GET /api/inventory/expiry-batches: every finished-product batch,
     soonest-expiry first, with computed status (EXPIRED / EXPIRING_SOON /
     OK) - the Inventory screen's new "Expiry Tracking" card.
   - Raw materials don't get a separate batch table - GRN already is a
     per-batch record (own price + DateReceived), so QC's raw-material
     batch list (/api/qc/grns) now carries the same computed expiry fields,
     keyed off DateReceived and the raw material's own name, so QC can
     check expiry alongside quality when deciding a batch.
   - Verified via curl against real product data: Kiri Packet 200ml
     (Milk Packets) produced 2026-09-07 -> expiry 2026-11-07; Set Yogurt Cup
     produced 2026-09-07 -> 2026-12-07; Ice Bar (standard) -> 2027-09-07 -
     all match the spec's worked examples exactly. Also verified FIFO
     depletion (Stock Out and a sale both correctly reduced batch
     quantities oldest-expiry-first) and EXPIRED/EXPIRING_SOON/OK status
     classification, then cleaned up every row of test data this
     introduced so the real database is back to its pre-test state.

+++ backend/src/main/java/erp_backend/controller/DashboardController.java
@@ -68,17 +68,17 @@ public class DashboardController {
         s.invoicesToday = invoices.stream().filter(i -> today.equals(i.getInvoiceDate())).count();
         s.salesTotalToday = invoices.stream()
                 .filter(i -> today.equals(i.getInvoiceDate()))
-                .mapToDouble(i -> i.getTotalAmount() != null ? i.getTotalAmount() : 0.0)
+                .mapToDouble(i -> i.getTotalAmount() != null ? i.getTotalAmount().doubleValue() : 0.0)
                 .sum();
         s.salesTotalThisMonth = invoices.stream()
                 .filter(i -> i.getInvoiceDate() != null
                         && i.getInvoiceDate().getMonthValue() == today.getMonthValue()
                         && i.getInvoiceDate().getYear() == today.getYear())
-                .mapToDouble(i -> i.getTotalAmount() != null ? i.getTotalAmount() : 0.0)
+                .mapToDouble(i -> i.getTotalAmount() != null ? i.getTotalAmount().doubleValue() : 0.0)
                 .sum();
 
         s.totalOutstandingBalance = customerRepository.findAll().stream()
-                .mapToDouble(c -> c.getOutstandingBalance() != null ? c.getOutstandingBalance() : 0.0)
+                .mapToDouble(c -> c.getOutstandingBalance() != null ? c.getOutstandingBalance().doubleValue() : 0.0)
                 .sum();
 
         List<FinishedProductStockMovement> movements = finishedProductStockMovementRepository.findAll();
